fix logging

+++ miranda/convert/_utils.py
@@ -240,13 +240,17 @@ def reanalysis_processing(
                             ]
 
                             jobs = list()
+                            if output_format != "zarr" and overwrite:
+                                logging.warning(
+                                    f"Removing existing {output_format} files for {var}."
+                                )
                             for i, d in enumerate(datasets):
-                                if out_filenames[i].exists() and overwrite:
-                                    logging.warning(
-                                        f"Removing existing {output_format} files for {var}."
-                                    )
-                                    if out_filenames[i].is_file():
-                                        out_filenames[i].unlink()
+                                if (
+                                    out_filenames[i].exists()
+                                    and out_filenames[i].is_file()
+                                    and overwrite
+                                ):
+                                    out_filenames[i].unlink()
 
                                 if (
                                     not out_filenames[i].exists()
